requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
boto3==1.34.56
botocore==1.34.56
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
jmespath==1.0.1
multidict==6.0.5
python-dateutil==2.9.0.post0
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
yarl==1.9.4
